A-Jar build and test jar file
