+++ data/duke.txt
@@ -1,3 +1,1 @@
-T,1,eat noodles
-D,0,do homework,by,2019-12-01
 T,0,eat food
