+++ src/main/java/duke/Deadline.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 public class Deadline extends Event {
