+++ src/main/java/duke/DukeException.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 public class DukeException extends Exception {
     DukeException(String errorMessage) {
