+++ src/main/java/duke/TaskList.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 import java.util.ArrayList;
