+++ src/main/java/duke/ToDo.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 public class ToDo extends Task{
     ToDo(String description) {
