+++ src/test/java/duke/ParserTest.java
@@ -1,4 +1,4 @@
-package Duke;
+package duke;
 
 import org.junit.jupiter.api.Test;
